+++ Server/Assets/Backend/Noisers/NoiserErlang.cs
@@ -6,24 +6,24 @@ namespace Assets.Backend.Noisers
 {
     public class NoiserErlang : Noiser
     {
-        public NoiserErlang(float _lambda, int _k)
+        public NoiserErlang(double _lambda, int _k)
         {
             lambda = _lambda;
             k = _k;
         }
 
-        public override float Get()
+        public override double Get()
         {
-            float result = 1.0f;
+            double result = 1.0;
 
             for (int i = 0; i < k; i++)
                 result *= getRandom();
 
-            result = -(float)Math.Log(result) / lambda;
+            result = -Math.Log(result) / lambda;
 
             return result;
         }
 
-        protected float lambda, k;
+        protected double lambda, k;
     }
 }
