+++ Server/Assets/Backend/Noisers/NoiserUniform.cs
@@ -5,17 +5,17 @@ namespace Assets.Backend.Noisers
 {
     public class NoiserUniform : Noiser
     {
-        public NoiserUniform(float _min, float _max)
+        public NoiserUniform(double _min, double _max)
         {
             min = _min;
             max = _max;
         }
 
-        public override float Get()
+        public override double Get()
         {
             return min + (max - min) * getRandom();
         }
 
-        protected float min, max;
+        protected double min, max;
     }
 }
